reduce raster pixel aliasing at high zoom levels
If the number of samples is limited to the number of raster pixels intersecting a tile, then at high zoom levels some neighboring pixels will both sample the same raster pixel, with this aliasing resulting in very visible doubled pixels. Increasing the maximum number of samples by a factor of three in each dimension reduces this effect as the doubled samples are a third of the raster pixel size. The only performance impact is at high zoom levels where doubled pixels would otherwise be visible, and performance is still no worse than at low zoom levels.