fully resolve raster pixels at high zoom levels
When zoomed in such that there are fewer raster pixels than the specified
resolution, draw raster pixels as rectangles with raster pixel boundaries
accurate to screen resolution if possible in the raster reference system.
Artifacts are visible if the raster pixels are not aligned with the axes
of the map/screen.  Avoids performance loss from over-sampling.